+++ src/main/java/org/example/NaiveOnArray/NaiveOnArray.java
@@ -1,7 +1,7 @@
 package org.example.NaiveOnArray;
 
 public class NaiveOnArray {
-    public static int[][] multiply(int[][] a, int[][] b) {
+    public static double[][] multiply(double[][] a, double[][] b) {
         int rowsA = a.length;
         int colsA = a[0].length;
         int rowsB = b.length;
@@ -11,7 +11,7 @@ public class NaiveOnArray {
             throw new IllegalArgumentException("Las matrices no se pueden multiplicar. El número de columnas de la primera matriz debe ser igual al número de filas de la segunda matriz.");
         }
 
-        int[][] c = new int[rowsA][colsB];
+        double[][] c = new double[rowsA][colsB];
         for (int i = 0; i < rowsA; i++) {
             for (int j = 0; j < colsB; j++) {
                 for (int k = 0; k < colsA; k++) {
